refactor: remove wrapper functions and use LaunchQLProject directly
- Delete deployModules, revertModules, verifyModules wrapper functions from migration.ts
- Remove MigrationOptions interface
- Update CLI commands (deploy, revert, verify) to use LaunchQLProject directly
- Update CoreDeployTestFixture to use LaunchQLProject methods
- Update exports in index.ts to only export getAvailableModules
- All tests passing (143 core + 7 cli tests)
